refactor: moved out supporting components
