+++ apps/journeys-admin/src/components/TemplateGallery/TagCarousels/CollectionButton/index.ts
@@ -1,0 +1,1 @@
+export { CollectionButton } from './CollectionButton'
