+++ apps/journeys-admin/src/components/TemplateGallery/TagCarousels/FeltNeedsButton/index.ts
@@ -1,0 +1,1 @@
+export { FeltNeedsButton } from './FeltNeedsButton'
